Cleaned code

+++ src/backend/components/deck/Hand.cpp
@@ -12,35 +12,35 @@ Hand::~Hand() {
 }
 
 std::unique_ptr<Card> Hand::popCard(uint8_t cardId) {
-    for(int i = 0; i < creatures.size(); i += 1)
+    for(size_t i = 0; i < creatures.size(); i += 1)
         if(creatures.at(i)->getCardUuid() == cardId) {
             std::unique_ptr<Card> card = std::move(creatures.at(i));
             creatures.erase(creatures.begin() + i);
             return std::move(card);
         }
             
-    for(int i = 0; i < lands.size(); i += 1)
+    for(size_t i = 0; i < lands.size(); i += 1)
         if(lands.at(i)->getCardUuid() == cardId) {
             std::unique_ptr<Card> card = std::move(lands.at(i));
             lands.erase(lands.begin() + i);
             return std::move(card);
         }
 
-    for(int i = 0; i < instants.size(); i += 1)
+    for(size_t i = 0; i < instants.size(); i += 1)
         if(instants.at(i)->getCardUuid() == cardId) {
             std::unique_ptr<Card> card = std::move(instants.at(i));
             instants.erase(instants.begin() + i);
             return std::move(card);
         }
         
-    for(int i = 0; i < enchantements.size(); i += 1)
+    for(size_t i = 0; i < enchantements.size(); i += 1)
         if(enchantements.at(i)->getCardUuid() == cardId) {
             std::unique_ptr<Card> card = std::move(enchantements.at(i));
             enchantements.erase(enchantements.begin() + i);
             return std::move(card);
         }
     
-    for(int i = 0; i < sorceries.size(); i += 1)
+    for(size_t i = 0; i < sorceries.size(); i += 1)
         if(sorceries.at(i)->getCardUuid() == cardId) {
             std::unique_ptr<Card> card = std::move(sorceries.at(i));
             sorceries.erase(sorceries.begin() + i);
@@ -158,7 +158,7 @@ std::vector<Creature*> Hand::getCreatures() {
 
 std::vector<Land*> Hand::getLands() {
     std::vector<Land*> result;
-    for(int i = 0; i < lands.size(); i += 1)
+    for(size_t i = 0; i < lands.size(); i += 1)
         result.push_back(lands.at(i).get());
     return result;
 }
